edits by comment

+++ src/card/card.service.ts
@@ -1,25 +1,38 @@
 import { Injectable } from '@nestjs/common';
-import { ICard } from '@interface/ICard';
 import { Card } from '@card/card.model';
 import { CardRepository } from '@card/card.repository';
 
 @Injectable()
-export class CardService implements ICard {
+export class CardService {
   constructor(private readonly cardRepository: CardRepository) {}
 
-  async create(dto: any): Promise<any> {
-    return this.cardRepository.create(dto);
+  async create(dto: any): Promise<Card> {
+    const card = await this.cardRepository.findByDevNumber(dto.devNumber);
+    if (!card) {
+      return this.cardRepository.create(Card.create(dto));
+    }
+    throw new Error('Error');
+  }
+  async findByDevNumber(dev: string): Promise<Card> {
+    const card = await this.cardRepository.findByDevNumber(dev);
+    if (card) {
+      return card;
+    }
+    throw new Error('Error');
+  }
+  async delete(dev: string): Promise<void> {
+    const card = await this.cardRepository.findByDevNumber(dev);
+    if (card) {
+      return this.cardRepository.delete(dev);
+    }
+    throw new Error('Error');
   }
 
-  async delete(dev: string): Promise<any> {
-    return this.cardRepository.delete(dev);
-  }
-
-  async findByDev(dev: string): Promise<any> {
-    return this.cardRepository.findByDev(dev);
-  }
-
-  async updateByDev(dev: string, dto: any): Promise<void> {
-    return this.cardRepository.updateByDev(dev, dto);
+  async update(dev: string, dto: any): Promise<Card> {
+    const card = await this.cardRepository.findByDevNumber(dev);
+    if (card) {
+      return this.cardRepository.update(dev, Card.create(dto));
+    }
+    throw new Error('Error');
   }
 }
